[refactor/slices] Slices cleanup
