+++ src/store/game/flow/flow.slice.ts
@@ -1,7 +1,7 @@
 import { EStatus } from "@/shared/enums/status.enum";
 import { IFlowSlice, TFlowSliceCreator } from "./flow.types";
 
-export const createFlowSlice: TFlowSliceCreator<IFlowSlice> = (set, get) => ({
+export const createFlowSlice: TFlowSliceCreator<IFlowSlice> = (set) => ({
   gameStatus: EStatus.SELECT_ACTION,
   round: 1,
