+++ src/store/game/players/tests/setup/create-test-store.ts
@@ -1,6 +1,5 @@
 import { create } from "zustand";
-import { IPlayersSlice } from "../../players.types";
-import { createPlayersSlice } from "../..";
+import { createPlayersSlice, IPlayersSlice } from "../..";
 
 export const createTestStore = () =>
   create<IPlayersSlice>((...args) => ({
